Move the code for fist-time configuration notice outside of workouts integration and update copy & link
