+++ src/integrations/admin/workouts-integration.php
@@ -45,13 +45,6 @@ class Workouts_Integration implements Integration_Interface {
 	private $product_helper;
 
 	/**
-	 * The indexing helper.
-	 *
-	 * @var Indexing_Helper
-	 */
-	private $indexing_helper;
-
-	/**
 	 * {@inheritDoc}
 	 */
 	public static function get_conditionals() {
@@ -65,20 +58,17 @@ class Workouts_Integration implements Integration_Interface {
 	 * @param WPSEO_Admin_Asset_Manager $admin_asset_manager The admin asset manager.
 	 * @param Options_Helper            $options_helper      The options helper.
 	 * @param Product_Helper            $product_helper      The product helper.
-	 * @param Indexing_Helper           $indexing_helper     The indexing helper.
 	 */
 	public function __construct(
 		WPSEO_Addon_Manager $addon_manager,
 		WPSEO_Admin_Asset_Manager $admin_asset_manager,
 		Options_Helper $options_helper,
-		Product_Helper $product_helper,
-		Indexing_Helper $indexing_helper
+		Product_Helper $product_helper
 	) {
 		$this->addon_manager       = $addon_manager;
 		$this->admin_asset_manager = $admin_asset_manager;
 		$this->options_helper      = $options_helper;
 		$this->product_helper      = $product_helper;
-		$this->indexing_helper     = $indexing_helper;
 	}
 
 	/**
@@ -88,7 +78,6 @@ class Workouts_Integration implements Integration_Interface {
 		\add_filter( 'wpseo_submenu_pages', [ $this, 'add_submenu_page' ], 8 );
 		\add_filter( 'wpseo_submenu_pages', [ $this, 'remove_old_submenu_page' ], 10 );
 		\add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_assets' ], 11 );
-		\add_action( 'admin_notices', [ $this, 'configuration_workout_notice' ] );
 	}
 
 	/**
@@ -185,85 +174,6 @@ class Workouts_Integration implements Integration_Interface {
 		}
 
 		echo '<div id="wpseo-workouts-container-free" class="yoast"></div>';
-	}
-
-	/**
-	 * Determines whether and where the "First-time SEO Configuration workout" admin notice should be displayed.
-	 *
-	 * @return bool Whether the "First-time SEO Configuration workout" admin notice should be displayed.
-	 */
-	public function should_display_configuration_workout_notice() {
-		if ( ! $this->options_helper->get( 'dismiss_configuration_workout_notice', false ) === false ) {
-			return false;
-		}
-
-		if ( ! $this->user_can_do_configuration_workout() ) {
-			return false;
-		}
-
-		if ( ! $this->on_wpseo_admin_page_or_dashboard() ) {
-			return false;
-		}
-
-		if ( $this->is_configuration_workout_finished() ) {
-			return false;
-		}
-
-		if ( $this->options_helper->get( 'first_time_install', false ) === false ) {
-			return false;
-		}
-
-		return ! $this->are_site_representation_name_and_logo_set() || $this->indexing_helper->get_unindexed_count() > 0;
-	}
-
-	/**
-	 * Displays an admin notice when the configuration workout has not been finished yet.
-	 *
-	 * @return void
-	 */
-	public function configuration_workout_notice() {
-		if ( ! $this->should_display_configuration_workout_notice() ) {
-			return;
-		}
-
-		$this->admin_asset_manager->enqueue_style( 'monorepo' );
-
-		$notice = new Notice_Presenter(
-			\__( 'First-time SEO configuration', 'wordpress-seo' ),
-			\sprintf(
-			/* translators: 1: Link start tag to the configuration workout, 2: Yoast SEO, 3: Link closing tag. */
-				\__( 'Get started quickly with the %1$s%2$s Configuration workout%3$s and configure Yoast SEO with the optimal SEO settings for your site!', 'wordpress-seo' ),
-				'<a href="' . \esc_url( \self_admin_url( 'admin.php?page=wpseo_workouts' ) ) . '">',
-				'Yoast SEO',
-				'</a>'
-			),
-			'mirrored_fit_bubble_woman_1_optim.svg',
-			null,
-			true,
-			'yoast-configuration-workout-notice'
-		);
-
-		//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- Output from present() is considered safe.
-		echo $notice->present();
-
-		// Enable permanently dismissing the notice.
-		echo "<script>
-			function dismiss_configuration_workout_notice(){
-				var data = {
-				'action': 'dismiss_configuration_workout_notice',
-				};
-
-				jQuery.post( ajaxurl, data, function( response ) {
-					jQuery( '#yoast-configuration-workout-notice' ).hide();
-				});
-			}
-
-			jQuery( document ).ready( function() {
-				jQuery( 'body' ).on( 'click', '#yoast-configuration-workout-notice .notice-dismiss', function() {
-					dismiss_configuration_workout_notice();
-				} );
-			} );
-			</script>";
 	}
 
 	/**
@@ -438,68 +348,5 @@ class Workouts_Integration implements Integration_Interface {
 	 */
 	private function user_can_edit_wordpress_options() {
 		return \current_user_can( 'manage_options' );
-	}
-
-	/**
-	 * Whether the user is currently visiting one of our admin pages or the WordPress dashboard.
-	 *
-	 * @return bool Whether the current page is a Yoast SEO admin page
-	 */
-	private function on_wpseo_admin_page_or_dashboard() {
-		$pagenow = $GLOBALS['pagenow'];
-
-		// Show on the WP Dashboard.
-		if ( $pagenow === 'index.php' ) {
-			return true;
-		}
-
-		$page_from_get = \filter_input( \INPUT_GET, 'page' );
-
-		// Show on Yoast SEO pages, with some exceptions.
-		if ( $pagenow === 'admin.php' && \strpos( $page_from_get, 'wpseo' ) === 0 ) {
-			$exceptions = [
-				'wpseo_workouts',
-				'wpseo_installation_successful',
-				'wpseo_installation_successful_free',
-			];
-
-			if ( ! \in_array( $page_from_get, $exceptions, true ) ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/**
-	 * Whether all steps of the configuration workout have been finished.
-	 *
-	 * @return bool Whether the configuration workout has been finished.
-	 */
-	private function is_configuration_workout_finished() {
-		$workouts_option = $this->options_helper->get( 'workouts_data', [] );
-
-		return \count( $workouts_option['configuration']['finishedSteps'] ) === 5;
-	}
-
-	/**
-	 * Whether the site representation name and logo have been set.
-	 *
-	 * @return bool  Whether the site representation name and logo have been set.
-	 */
-	private function are_site_representation_name_and_logo_set() {
-		$company_or_person = $this->options_helper->get( 'company_or_person', '' );
-
-		if ( $company_or_person === '' ) {
-			return false;
-		}
-
-		if ( $company_or_person === 'company' ) {
-			return ! empty( $this->options_helper->get( 'company_name' ) )
-				&& ! empty( $this->options_helper->get( 'company_logo', '' ) );
-		}
-
-		return ! empty( $this->options_helper->get( 'company_or_person_user_id' ) )
-			&& ! empty( $this->options_helper->get( 'person_logo', '' ) );
 	}
 }
